+++ src/constants/index.js
@@ -1,0 +1,13 @@
+export const DISPENSER_TIME_INTERVAL = 500;
+
+export const INSERT_MONEY_RANGE = {
+  MIN: 100,
+  MAX: 50000,
+};
+
+export const DRINK_PRICE = {
+  milk: 1000,
+  espresso: 1000,
+  americano: 1500,
+  cafeLatte: 2000,
+};
